P4C3 modify route DELETE

+++ controllers/stuff.js
@@ -1,4 +1,5 @@
 const Thing = require('../models/thing');
+const fs = require('fs')
 
 exports.createThing = (req, res, next) => {
     const thingObject = JSON.parse(req.body.thing);
@@ -44,11 +45,24 @@ exports.modifyThing = (req, res, next) => {
         });
 };
 
-exports.deleteThing = ((req, res, next) => {
-    Thing.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Objet supprimé !' }))
-        .catch(error => res.status(400).json({ error }));
-});
+exports.deleteThing = (req, res, next) => {
+    Thing.findOne({ _id: req.params.id })
+        .then(thing => {
+            if (thing.userId != req.auth.userId) {
+                res.status(401).json({ message: 'Not authorized' });
+            } else {
+                const filename = thing.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Thing.deleteOne({ _id: req.params.id })
+                        .then(() => { res.status(200).json({ message: 'Objet supprimé !' }) })
+                        .catch(error => res.status(401).json({ error }));
+                });
+            }
+        })
+        .catch(error => {
+            res.status(500).json({ error });
+        });
+};
 
 exports.getAllThings = ((req, res, next) => {
     Thing.find()
@@ -58,6 +72,6 @@ exports.getAllThings = ((req, res, next) => {
 
 exports.deleteAllThings = ((req, res, next) => {
     Thing.deleteMany()
-        .then(things => res.status(200).json({ message: 'Tous les Objets supprimés !' }))
+        .then(() => res.status(200).json({ message: 'Tous les Objets supprimés !' }))
         .catch(error => res.status(400).json({ error }));
 });
